Extract tests for isUint32Array to own file
